Add separated files
