+++ Buform/Platforms/Ios/Groups/Text/DefaultFormGroupHandler.cs
@@ -1,0 +1,8 @@
+using Buform.Groups;
+
+namespace Buform;
+
+public class DefaultFormGroupHandler : FormGroupHandlerBase<IFormGroup>
+{
+    /* Nothing to do */
+}
